% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteDFLEtable.R
\name{CompleteDFLEtable}
\alias{CompleteDFLEtable}
\title{Complete a dataframe with variables useful for the calculation of disability-free life expectancies (DFLE)}
\usage{
CompleteDFLEtable(tab)
}
\arguments{
\item{tab}{a dataframe containing some of the variables}
}
\value{
a dataframe with all variables that can be calculated from the input dataframe
}
\description{
Take an input dataset containing some information on age, mortality rates and prevalences
of disability (at least 'age', 'qx' or 'mx', and 'pix' must be within the input dataset)
and complete it with  variables for life expectancies and disability-free life expectancies.
#' If some indispensable variables are missing (eg 'age'), the output is the same as the input dataset.
}
\details{
If variables 'sex' and/or 'year' are in the input dataset, calculations are performed for each
separate values of 'sex' and/or 'year'.

Calculations follow example 1 in the sullivan manual (June 2007 version).
See https://reves.site.ined.fr/en/resources/computation_online/sullivan/ for more explanations.
}
\examples{
CompleteDFLEtable( sullivan[,c("year","age","mx","qx","pix")] )
CompleteDFLEtable( FRInseeMortalityForecast2016 \%>\% mutate(pix = (age/150)^2, mx = qx) \%>\% filter(year \%in\% c(2013,2015,2020,2030)) )
}
