% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{description_sullivan}
\alias{description_sullivan}
\title{A description of the variables in the dataset from Sullivan manual's calculation example.}
\format{
A data frame with 2 variables:
\describe{
\item{heading}{name of the variable in the sullivan dataset}
\item{description}{description of the variable from the sullivan dataset}
}
}
\source{
\url{https://reves.site.ined.fr/fichier/s_rubrique/20182/sullivan_manual_jun2007.en.xls}
}
\usage{
description_sullivan
}
\description{
A description of the variables in the dataset from Sullivan manual's calculation example.
}
\keyword{datasets}
