% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRGaliEUSilc}
\alias{FRGaliEUSilc}
\title{A dataset with prevalence of general activity limitations (GALI) in France}
\format{
A data frame with 1248 observations and 5 variables:
\describe{
\item{incap}{type of general activity limitations ('gali_incl_moderate' or 'gali_severe')}
\item{prevalence}{share of people with general activity limitations in the given age bracket (number between 0 and 1)}
\item{sex}{sex ('female' or 'male')}
\item{agebracket}{age bracket (eg '[60;65)' = between 60 and 64 years old)}
\item{year}{year of observation (value for the month of December)}
}
}
\source{
\url{https://drees.solidarites-sante.gouv.fr/publications/etudes-et-resultats/en-2020-lesperance-de-vie-sans-incapacite-65-ans-est-de-121-ans}
}
\usage{
FRGaliEUSilc
}
\description{
Prevalences are measured in the SRCV survey ie. the French version of EU-SILC.
They are measured by sex and age bracket, for severe general activity limitations and for limitations including moderate ones.
}
\keyword{datasets}
