% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRInseeMortalityForecast2021}
\alias{FRInseeMortalityForecast2021}
\title{A dataset with mortality rates from Insee's 2021 population forecast (central scenario)}
\format{
A data frame with 115434 observations and 6 variables:
\describe{
\item{age}{age}
\item{year}{year}
\item{sex}{sex ('male' or 'female' or 'all')}
\item{type.obs}{type of observation ('observed' or 'observed (prov.)' or 'forecasted')}
\item{def.age}{definition of age ('age at end of year' or 'current age (approx)')}
\item{qx}{mortality rates at each year, age, sex}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/5894093?sommaire=5760764}
}
\usage{
FRInseeMortalityForecast2021
}
\description{
The dataset includes observed values from 1962 to 2020, then forecasted values from 2021 to 2120.
}
\details{
Note: Insee's file provides number of death by age at the end of the year, not age at last birthday.
We assume deaths are evenly distributed within the year, so we use simple average to convert from
definition of age to the other.
}
\keyword{datasets}
