% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRInseePopulationForecast2021}
\alias{FRInseePopulationForecast2021}
\title{A dataset with population by sex and age, from Insee's 2021 population forecast (central scenario)}
\format{
A data frame with 50325 observations and 6 variables:
\describe{
\item{age0101}{age at January, 1st of the year}
\item{year}{year}
\item{sex}{sex ('male' or 'female' or 'all')}
\item{geo}{geographical field ('france' only)}
\item{type.obs}{'observed' or 'forecasted'}
\item{popx0101}{population on the 1st of January at each year, age, sex}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/5894093?sommaire=5760764}
}
\usage{
FRInseePopulationForecast2021
}
\description{
The dataset also includes observed data from 1990 to 2022. Population is measured on the 1st of January of each year.
}
\keyword{datasets}
