% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRmortalityForecast2016}
\alias{FRmortalityForecast2016}
\title{A dataset with mortality rates from Insee's 2016 population forecast (central scenario)}
\format{
A data frame with 14036 observations and 4 variables:
\describe{
\item{age}{age}
\item{year}{year}
\item{sex}{sex ('male' or 'female')}
\item{qx}{mortality rates at each year, age, sex}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2496716?sommaire=2496793}
}
\usage{
FRmortalityForecast2016
}
\description{
Note: Insee's file provides number of death by age at the end of the year, not age at last birthday.
We assume deaths are evenly distributed within the year, so we use simple average to convert from
definition of age to the other.
}
\keyword{datasets}
