% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_to_polynomial.R
\name{prevalence_to_polynomial}
\alias{prevalence_to_polynomial}
\title{Estimates a polynomial approximation for prevalences by age}
\usage{
prevalence_to_polynomial(age, prevalence, weight = rep(1, NROW(prevalence)))
}
\arguments{
\item{age}{a vector with observed ages}

\item{prevalence}{a vector with observed prevalences}

\item{weight}{a vector of weights for the regression (optional)}
}
\value{
a vector with prevalences according to polynomial approximation
}
\description{
Given a vector of ages and a vector of prevalences, the function returns the coefficients
of a polynomial approximation of degree 4
}
\details{
Prevalences are supposed to be constant for all ages superior to 100
}
