% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalenceForecast.R
\name{prevalenceForecast}
\alias{prevalenceForecast}
\title{Forecasts prevalences of disability and DFLEs}
\usage{
prevalenceForecast(tabref, tabproj, hypo, includevars = c())
}
\arguments{
\item{tabref}{a data frame for the reference year, containing variables: year, age, qx (mortality rate) and pix (prevalences) (+ optional: sex)}

\item{tabproj}{a data frame for projection years, containing variables: year, age, and qx (mortality rate) (+optional: sex)}

\item{hypo}{projection hypotheses ('cstDFLE','cstDLE','cstPreval','cstPctDFLE')}

\item{includevars}{a vector of names of variables that must be included in the output table (otherwise, only pix, ex, DFLEx, DLEx and pctDFLEx are kept)}
}
\value{
a data frame with prevalences, DFLE, DLE and share of DFLE in total LE, by year and age (and optionnally sex)
}
\description{
Given a table for the reference year with mortality rates and prevalences of disability, and
a projection table with mortality rates, the function calculates prevalences and disability-free
life expectancies.
}
\details{
Projection hypotheses can be one of the following:
\itemize{
\item 'cstDFLE' : disability-free life expectancies constant in all year
\item 'cstDLE' : in-disability life expectancies constant in all year
\item 'cstPreval' : prevalences by age (and sex) constant in all year
\item 'cstPctDFLE' : share of disability-free life expectancy in total life-expectancy constant
}
}
