% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Runs examples of uses of the package}
\usage{
runExample(example = "projprevalence")
}
\arguments{
\item{example}{name of the example (default is 'projprevalence' : a Shiny app displaying forecasting of prevalences and DFLE for France)}
}
\value{
launches the app
}
\description{
Runs examples of uses of the package
}
