% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sullivan}
\alias{sullivan}
\title{An example of calculation using Sullivan's method.}
\format{
A data frame with  rows and  variables:
\describe{
\item{year}{year}
}
}
\source{
\url{https://reves.site.ined.fr/fichier/s_rubrique/20182/sullivan_manual_jun2007.en.xls}
}
\usage{
sullivan
}
\description{
A dataset taken from the 'Ex 1' example in the Sullivan manual (june 2007 version)
available at https://reves.site.ined.fr/en/resources/computation_online/sullivan/
}
\keyword{datasets}
