% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRDreesAPA}
\alias{FRDreesAPA}
\title{A dataset with share of APA beneficiaries in France}
\format{
A data frame with 138 observations and 6 variables:
\describe{
\item{typepresta}{type of benefits (eg. 'à domicile' or 'en établissement')}
\item{prevalence}{share of beneficiaires within the total population in the given age bracket (number between 0 and 1)}
\item{sex}{sex ('female' or 'male')}
\item{agebracket}{age bracket (eg '[60;65)' = between 60 and 64 years old)}
\item{year}{year of observation (value for the monthe of December)}
}
}
\source{
\url{https://drees.solidarites-sante.gouv.fr/etudes-et-statistiques/publications/les-dossiers-de-la-drees/article/incapacites-et-perte-d-autonomie-des-personnes-agees-en-france-une-evolution}
}
\usage{
FRDreesAPA
}
\description{
APA is a benefit covering long term care expenditures both at home ('à domicile') or in institutions ('en établissement').
Shares of beneficiaries are estimated through the 2016, 2017 and 2018 'Aide sociale' survey.
They are estimated by age brackets (eg. '[60;65)' stands for ages 60 to 64)
A more complete dataset of APA prevalences, as well as prevalences for other social allowances in France, can be found in the 'asdep' package.
}
\keyword{datasets}
