% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRDreesEHPA}
\alias{FRDreesEHPA}
\title{A dataset with proportions on people living in institutions in France according to Drees's EHPA survey}
\format{
A data frame with 981 observations and 6 variables:
\describe{
\item{year}{year of observation}
\item{sex}{sex ('female' or 'male' or 'all)}
\item{age}{age}
\item{nb}{number of elder people living in institutions}
\item{poptot}{total population by age and sexe}
\item{prevalence}{share of elder people living in institutions}
}
}
\source{
\url{https://drees.solidarites-sante.gouv.fr/publications-communique-de-presse/etudes-et-resultats/251104_ER_etablissements-hebergement-personnes-agees}
}
\usage{
FRDreesEHPA
}
\description{
The EHPA survey is collected every four years. Data are available for 2007, 2011, 2015, 2019, and 2023. Numbers of people are measured on the Decembre 31st of each year.
}
\keyword{datasets}
