% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRInseeMortalityrates_t69}
\alias{FRInseeMortalityrates_t69}
\title{An (alternative) dataset with observed mortality rates by sex and age, from Insee}
\format{
A data frame with 67872 observations and 7 variables:
\describe{
\item{age}{age}
\item{def.age}{definition of age: at the end of the year vs. current age}
\item{year}{year}
\item{sex}{sex ('male' or 'female' or 'all')}
\item{geo}{geographical field ('france' or 'metropolitan france')}
\item{status}{'p' for 'provisional', '' otherwise}
\item{qx}{mortality rate at each year, age, sex}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/fichier/8721189/}
}
\usage{
FRInseeMortalityrates_t69
}
\description{
This dataset corresponds to data released by Insee as T69, as comparison to the FRInseeMortalityrates, which
corresponds to T68 (see Insee's website for a presentation of the differences).
}
\details{
The current version of the data table was released in January, 2026 and contains observations
from 1946 (metropolitan france) or 1994 (france) to 2025. The geographical field is France including oversea territories.
}
\keyword{datasets}
